feat(data-s): add main data structures and test env (queue, vector, linked-list)
created the main data-structures with their own namespace that could be useful on the project. Furthermore, i added a test framework that was usefull to test some general features of the data-s. The test framework is visible in the makefile, where there is also a comment to describe how to run only a single file.