fix(vector|segment): vector resize was a reserve
fixed some naughty bugs, the resize was a reserve because it did't
update the size_. There was also a bug in the level costructor, because
the lenght_ was not updated when introducing a doorsegment.
Maybe we should make another costructor that takes the two ends in consideration
and more the segment with the + at the beginning is not intuitive.
its not intuitive....