feat(loaders): directory-loader first impl
created and integrated with both level/level.hpp and manager/level.hpp
the directory-loader and loader handles

POSSIBLE PROBLEMS:
the loader-handler.hpp module needs to have strict coesion of the file
rules, its easy to make mistakes about that, we should have some checks

The problems we have if the input file in assets is not well formed
are very weird memory mis-alignment errors, which are quite hard to debug.
So handle with care.